js updates

+++ controllers/clothingItems.js
@@ -3,6 +3,7 @@ const { ValidationError } = require("../utils/errors/ValidationError");
 const { NotFoundError } = require("../utils/errors/NotFoundError");
 const { CastError } = require("../utils/errors/CastError");
 const { ServerError } = require("../utils/errors/ServerError");
+const { ForbiddenError } = require("../utils/errors/ForbiddenError");
 
 const createItem = (req, res) => {
   // console.log(req);
@@ -50,36 +51,42 @@ const deleteItem = (req, res) => {
   const { itemId } = req.params;
   console.log(itemId);
 
-  ClothingItem.findByIdAndDelete(itemId)
+  ClothingItem.findById(itemId)
     .orFail(() => new NotFoundError())
-    .then(() =>
-      res
-        .status(200)
-        .send({ message: "item deleted" })
-    ).catch((e) => {
-      if (e.name === "CastError") {
-        const castError = new CastError();
+    .then((item) => {
+      if (item.owner.id !== item.id) {
+        const forbiddenError = new ForbiddenError();
         return res
-          .status(castError.statusCode)
-          .send({ message: castError.message });
+          .status(forbiddenError.statusCode)
+          .res.send({ message: forbiddenError.message });
+      } else {
+        return ClothingItem.findByIdAndDelete(itemId)
+          .orFail(() => new NotFoundError())
+          .then(() => res.status(200).send({ message: "item deleted" }))
+          .catch((e) => {
+            if (e.name === "CastError") {
+              const castError = new CastError();
+              return res
+                .status(castError.statusCode)
+                .send({ message: castError.message });
+            }
+            if (e.name && e.name === "NotFoundError") {
+              const notFoundError = new NotFoundError();
+              return res
+                .status(notFoundError.statusCode)
+                .send({ message: notFoundError.message });
+            }
+            const serverError = new ServerError();
+            return res
+              .status(serverError.statusCode)
+              .send({ message: serverError.message });
+          });
       }
-      if (e.name && e.name === "NotFoundError") {
-        const notFoundError = new NotFoundError();
-        return res
-          .status(notFoundError.statusCode)
-          .send({ message: notFoundError.message });
-      }
-        const serverError = new ServerError();
-        return res
-          .status(serverError.statusCode)
-          .send({ message: serverError.message });
     });
-  }
+};
 
 module.exports = {
   createItem,
   getItems,
   deleteItem,
 };
-
-
